SteamTarget: Enable UIPI
In order to stay on top of UWP Apps, UAC Bypass protection has to be enabled.
Unfortunately this also means having to have a signed executable and that executable being placed inside a trusted location ("C:\Program Files\", "C:\Program Files (x86)", ...) for the Program to run.

Enabled in Release type builds only