Support passing command line arguments to targets (closes #6)
This also refactors target configuration storage to reduce the redundancy in copying the same set of files for every target. Each configured entry now instead uses a .ini file with the name of the entry stored in ./targets. SteamTarget can accept the path to a .ini file as a command line argument.

This also fixes the UI so that inputs are ordered the same as the visual order so that tab behavior isn't confusing.