Filter and sort UWP apps; fixes missing logos
Filters out apps stored in SystemApps, or Microsoft apps that begin with the name "Windows", or language packs, or that match the phrase "Sysinternals", or have bad display names and are produced by Microsoft.
For missing logos, find the largest variant of the asset and use that. Works around the use of qualifiers in file names, which Windows automatically handles, but which we can't.
Sort the app list before passing it back to the GlosSI app.